requirements.txt
asttokens==2.0.7
backcall==0.2.0
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.9.1
Flask==2.2.0
google-auth==2.10.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
kubernetes==24.2.0
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.3.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.2.1
